Added Login activity and relevant icons to app home screen
Created a login activity, but we shouldn't worry about this now. It doesn't affect anything at the moment, and we will need to make a way for the user to create an account before this will work well.

Changed bottom of main activity to have tabs for home, records, and plots. Also added relevant icons to these parts